+++ recipe/emailpassword/api/generatePasswordResetToken.go
@@ -44,11 +44,15 @@ func GeneratePasswordResetToken(apiImplementation epmodels.APIInterface, options
 		return err
 	}
 
-	_, err = (*apiImplementation.GeneratePasswordResetTokenPOST)(formFields, options, &map[string]interface{}{})
+	resp, err := (*apiImplementation.GeneratePasswordResetTokenPOST)(formFields, options, &map[string]interface{}{})
 	if err != nil {
 		return err
 	}
-	return supertokens.Send200Response(options.Res, map[string]interface{}{
-		"status": "OK",
-	})
+	if resp.OK != nil {
+		return supertokens.Send200Response(options.Res, map[string]interface{}{
+			"status": "OK",
+		})
+	} else {
+		return supertokens.Send200Response(options.Res, supertokens.ConvertGeneralErrorToJsonResponse(*resp.GeneralError))
+	}
 }