+++ recipe/emailverification/api/implementation.go
@@ -25,8 +25,20 @@ import (
 )
 
 func MakeAPIImplementation() evmodels.APIInterface {
-	verifyEmailPOST := func(token string, options evmodels.APIOptions, userContext supertokens.UserContext) (evmodels.VerifyEmailUsingTokenResponse, error) {
-		return (*options.RecipeImplementation.VerifyEmailUsingToken)(token, userContext)
+	verifyEmailPOST := func(token string, options evmodels.APIOptions, userContext supertokens.UserContext) (evmodels.VerifyEmailPOSTResponse, error) {
+		resp, err := (*options.RecipeImplementation.VerifyEmailUsingToken)(token, userContext)
+		if err != nil {
+			return evmodels.VerifyEmailPOSTResponse{}, err
+		}
+		if resp.OK != nil {
+			return evmodels.VerifyEmailPOSTResponse{
+				OK: resp.OK,
+			}, err
+		} else {
+			return evmodels.VerifyEmailPOSTResponse{
+				EmailVerificationInvalidTokenError: resp.EmailVerificationInvalidTokenError,
+			}, nil
+		}
 	}
 
 	isEmailVerifiedGET := func(options evmodels.APIOptions, userContext supertokens.UserContext) (evmodels.IsEmailVerifiedGETResponse, error) {