+++ recipe/thirdpartypasswordless/tplmodels/apiInterface.go
@@ -55,9 +55,7 @@ type ConsumeCodePOSTResponse struct {
 		MaximumCodeInputAttempts    int
 	}
 	RestartFlowError *struct{}
-	GeneralError     *struct {
-		Message string
-	}
+	GeneralError     *supertokens.GeneralErrorResponse
 }
 
 type ThirdPartySignInUpOutput struct {
@@ -68,5 +66,5 @@ type ThirdPartySignInUpOutput struct {
 		Session          sessmodels.SessionContainer
 	}
 	NoEmailGivenByProviderError *struct{}
-	FieldError                  *struct{ ErrorMsg string }
+	GeneralError                *supertokens.GeneralErrorResponse
 }